refactor: playerNameDTO, playerBettingMoneyDTO -> playerDTO 로 통합
