--- src/main/java/dto/RequestPlayerNameDTO.java
@@ -1,13 +1,0 @@
-package dto;
-
-public class RequestPlayerNameDTO {
-    private String name;
-
-    public RequestPlayerNameDTO(String name) {
-        this.name = name;
-    }
-
-    public String getName() {
-        return this.name;
-    }
-}
